+++ src/app/Scenes/Onboarding/Auth2/AuthScenes.tsx
@@ -1,18 +1,24 @@
 import { AuthScreen } from "app/Scenes/Onboarding/Auth2/components/AuthScreen"
-import { EmailSocialStep } from "app/Scenes/Onboarding/Auth2/scenes/EmailSocialStep"
+// import { EmailSocialStep } from "app/Scenes/Onboarding/Auth2/scenes/EmailSocialStep"
 import { ForgotPasswordStep } from "app/Scenes/Onboarding/Auth2/scenes/ForgotPasswordStep"
+import { LoginEmailStep } from "app/Scenes/Onboarding/Auth2/scenes/LoginEmailStep"
 import { LoginOTPStep } from "app/Scenes/Onboarding/Auth2/scenes/LoginOTPStep"
 import { LoginPasswordStep } from "app/Scenes/Onboarding/Auth2/scenes/LoginPasswordStep"
 import { SignUpNameStep } from "app/Scenes/Onboarding/Auth2/scenes/SignUpNameStep"
 import { SignUpPasswordStep } from "app/Scenes/Onboarding/Auth2/scenes/SignUpPasswordStep"
+import { WelcomeStep } from "app/Scenes/Onboarding/Auth2/scenes/WelcomeStep"
 import { ScrollView } from "react-native-gesture-handler"
 
 export const AuthScenes: React.FC = () => {
   return (
     <>
-      <ScrollView keyboardShouldPersistTaps>
-        <AuthScreen name="EmailSocialStep">
-          <EmailSocialStep />
+      <ScrollView keyboardShouldPersistTaps="always">
+        <AuthScreen name="WelcomeStep">
+          <WelcomeStep />
+        </AuthScreen>
+
+        <AuthScreen name="LoginEmailStep">
+          <LoginEmailStep />
         </AuthScreen>
 
         <AuthScreen name="LoginPasswordStep">
