+++ src/app/Scenes/Onboarding/Auth2/components/AuthModal.tsx
@@ -3,36 +3,58 @@ import { AuthContext } from "app/Scenes/Onboarding/Auth2/AuthContext"
 import { MotiView } from "moti"
 import { Dimensions, KeyboardAvoidingView, Platform } from "react-native"
 import { Easing } from "react-native-reanimated"
-import { useSafeAreaInsets } from "react-native-safe-area-context"
 
 const HEIGHT = {
-  EmailSocialStep: 280,
-  LoginPasswordStep: 300,
+  WelcomeStep: 300,
+  LoginEmailStep: 280,
+  LoginPasswordStep: 320,
+  ForgotPasswordStep: 320,
+  LoginOTPStep: 280,
+  OnboardingWebView: 300,
+  SignUpPasswordStep: 300,
+  SignUpNameStep: 500,
   collapsed: 300,
 }
 
 export const AuthModal: React.FC = ({ children }) => {
   const { isModalExpanded, isMounted, currentScreen } = AuthContext.useStoreState((state) => state)
 
-  const screenHeight = Dimensions.get("window").height
   const { space } = useTheme()
-  const insets = useSafeAreaInsets()
+
+  const screenHeight = Dimensions.get("window").height
+
+  const height = (() => {
+    if (isModalExpanded) {
+      return HEIGHT[currentScreen?.name ?? "WelcomeStep"]
+    }
+
+    return HEIGHT.collapsed
+  })()
+
+  const top = (() => {
+    // Position the modal in the center of the screen, minus some padding
+    if (isModalExpanded) {
+      return screenHeight / 2 - height / 2 - space(12)
+    }
+
+    // Position at the bottom of the screen, minus height of modal and padding
+    return screenHeight - height - space(6)
+  })()
 
   return (
     <KeyboardAvoidingView
       style={{ flex: 1 }}
       behavior={Platform.OS === "ios" ? "padding" : "height"}
     >
-      <Box justifyContent="flex-end" flex={1}>
+      <Box flex={1}>
         <MotiView
           from={{
             height: HEIGHT.collapsed,
-            bottom: space(4),
-            translateY: isModalExpanded ? 0 : screenHeight * 4,
+            top: isModalExpanded ? 0 : screenHeight,
           }}
           animate={{
-            height: isModalExpanded ? HEIGHT[currentScreen.name] : HEIGHT.collapsed,
-            translateY: 0,
+            height,
+            top,
           }}
           transition={{
             type: "timing",
@@ -45,6 +67,7 @@ export const AuthModal: React.FC = ({ children }) => {
             backgroundColor: "white",
             borderRadius: space(2),
             overflow: "hidden",
+            position: "relative",
           }}
         >
           <Flex height="100%" justifyContent="center" p={1}>
